+++ changes.md
@@ -8,6 +8,7 @@
   be called when an item is dropped.
 - Form with pat-autosubmit doesn't get notified when injected inputs change.
 - Inject pattern with `autoload-visible` injected erroneously based upon old autoload element no longer in DOM.
+- Add the class `modal-active` to the `body` element whenever a modal is in the DOM.
 
 ## 2.0.12 - Oct. 9, 2015